JUnits tuning. Add Database#isPeculiarLiquibaseSchema method to define whether a database support own schema name for liquibase tables.
git-svn-id: http://liquibase.jira.com/svn/CORE/trunk@910 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ core/src/java-test/liquibase/statementexecute/AbstractExecuteTest.java
@@ -146,28 +146,32 @@ public abstract class AbstractExecuteTest {
             if (database.supportsSchemas()) {
                 database.dropDatabaseObjects(TestContext.ALT_SCHEMA);
                 connection.commit();
-                try {
-                    connectionStatement.executeUpdate("drop table " + database.escapeTableName(TestContext.ALT_SCHEMA, database.getDatabaseChangeLogLockTableName()));
-                } catch (SQLException e) {
-                    ;
+                if (!database.isPeculiarLiquibaseSchema()) {
+                    try {
+                        connectionStatement.executeUpdate("drop table " + database.escapeTableName(TestContext.ALT_SCHEMA, database.getDatabaseChangeLogLockTableName()));
+                    } catch (SQLException e) {
+                        ;
+                    }
+                    connection.commit();
                 }
-                connection.commit();
-                try {
-                    connectionStatement.executeUpdate("drop table " + database.escapeTableName(TestContext.ALT_SCHEMA, database.getDatabaseChangeLogTableName()));
-                } catch (SQLException e) {
-                    ;
+                if (!database.isPeculiarLiquibaseSchema()) {
+                    try {
+                        connectionStatement.executeUpdate("drop table " + database.escapeTableName(TestContext.ALT_SCHEMA, database.getDatabaseChangeLogTableName()));
+                    } catch (SQLException e) {
+                        ;
+                    }
+                    connection.commit();
                 }
-                connection.commit();
             }
-            database.dropDatabaseObjects(null);
+            database.dropDatabaseObjects(database.convertRequestedSchemaToSchema(null));
             try {
-                connectionStatement.executeUpdate("drop table " + database.escapeTableName(null, database.getDatabaseChangeLogLockTableName()));
+                connectionStatement.executeUpdate("drop table " + database.escapeTableName(database.getLiquibaseSchemaName(), database.getDatabaseChangeLogLockTableName()));
             } catch (SQLException e) {
                 ;
             }
             connection.commit();
             try {
-                connectionStatement.executeUpdate("drop table " + database.escapeTableName(null, database.getDatabaseChangeLogTableName()));
+                connectionStatement.executeUpdate("drop table " + database.escapeTableName(database.getLiquibaseSchemaName(), database.getDatabaseChangeLogTableName()));
             } catch (SQLException e) {
                 ;
             }
